refactored certification and education
